Route with middleware

+++ src/routes/routes.php
@@ -41,7 +41,7 @@ Route::group(['prefix' => 'wilayah-indonesia'], function() {
 
 });
 
-Route::group(['prefix' => 'api/wilayah-indonesia/province', 'middleware' => ['web']], function() {
+Route::group(['prefix' => 'api/wilayah-indonesia/province', 'middleware' => ['auth', 'role:superadministrator|admin_sekolah']], function() {
     $class          = 'Bantenprov\WilayahIndonesia\Http\Controllers\ProvinceController';
     $name           = 'wilayah-indonesia.province';
     $controllers    = (object) [
@@ -55,17 +55,17 @@ Route::group(['prefix' => 'api/wilayah-indonesia/province', 'middleware' => ['we
         'destroy'   => $class.'@destroy',
     ];
 
-    Route::get('/',             $controllers->index)->name($name.'.index');
-    Route::get('/get',          $controllers->get)->name($name.'.get');
-    Route::get('/create',       $controllers->create)->name($name.'.create');
-    Route::get('/{id}',         $controllers->show)->name($name.'.show');
-    Route::post('/',            $controllers->store)->name($name.'.store');
-    Route::get('/{id}/edit',    $controllers->edit)->name($name.'.edit');
-    Route::put('/{id}',         $controllers->update)->name($name.'.update');
-    Route::delete('/{id}',      $controllers->destroy)->name($name.'.destroy');
+    Route::get('/',             $controllers->index)->name($name.'.index')->middleware(['role:superadministrator']);
+    Route::get('/get',          $controllers->get)->name($name.'.get')->middleware(['role:superadministrator|admin_sekolah']);
+    Route::get('/create',       $controllers->create)->name($name.'.create')->middleware(['role:superadministrator']);
+    Route::get('/{id}',         $controllers->show)->name($name.'.show')->middleware(['role:superadministrator']);
+    Route::post('/',            $controllers->store)->name($name.'.store')->middleware(['role:superadministrator']);
+    Route::get('/{id}/edit',    $controllers->edit)->name($name.'.edit')->middleware(['role:superadministrator']);
+    Route::put('/{id}',         $controllers->update)->name($name.'.update')->middleware(['role:superadministrator']);
+    Route::delete('/{id}',      $controllers->destroy)->name($name.'.destroy')->middleware(['role:superadministrator']);
 });
 
-Route::group(['prefix' => 'api/wilayah-indonesia/city', 'middleware' => ['web']], function() {
+Route::group(['prefix' => 'api/wilayah-indonesia/city', 'middleware' => ['auth', 'role:superadministrator|admin_sekolah']], function() {
     $class          = 'Bantenprov\WilayahIndonesia\Http\Controllers\CityController';
     $name           = 'wilayah-indonesia.city';
     $controllers    = (object) [
@@ -80,18 +80,18 @@ Route::group(['prefix' => 'api/wilayah-indonesia/city', 'middleware' => ['web']]
         'destroy'       => $class.'@destroy',
     ];
 
-    Route::get('/',                     $controllers->index)->name($name.'.index');
-    Route::get('/get',                  $controllers->get)->name($name.'.get');
-    Route::get('/get/by-province/{id}', $controllers->getByProvince)->name($name.'.get-by-province');
-    Route::get('/create',               $controllers->create)->name($name.'.create');
-    Route::get('/{id}',                 $controllers->show)->name($name.'.show');
-    Route::post('/',                    $controllers->store)->name($name.'.store');
-    Route::get('/{id}/edit',            $controllers->edit)->name($name.'.edit');
-    Route::put('/{id}',                 $controllers->update)->name($name.'.update');
-    Route::delete('/{id}',              $controllers->destroy)->name($name.'.destroy');
+    Route::get('/',                     $controllers->index)->name($name.'.index')->middleware(['role:superadministrator']);
+    Route::get('/get',                  $controllers->get)->name($name.'.get')->middleware(['role:superadministrator']);
+    Route::get('/get/by-province/{id}', $controllers->getByProvince)->name($name.'.get-by-province')->middleware(['role:superadministrator|admin_sekolah']);
+    Route::get('/create',               $controllers->create)->name($name.'.create')->middleware(['role:superadministrator']);
+    Route::get('/{id}',                 $controllers->show)->name($name.'.show')->middleware(['role:superadministrator']);
+    Route::post('/',                    $controllers->store)->name($name.'.store')->middleware(['role:superadministrator']);
+    Route::get('/{id}/edit',            $controllers->edit)->name($name.'.edit')->middleware(['role:superadministrator']);
+    Route::put('/{id}',                 $controllers->update)->name($name.'.update')->middleware(['role:superadministrator']);
+    Route::delete('/{id}',              $controllers->destroy)->name($name.'.destroy')->middleware(['role:superadministrator']);
 });
 
-Route::group(['prefix' => 'api/wilayah-indonesia/district', 'middleware' => ['web']], function() {
+Route::group(['prefix' => 'api/wilayah-indonesia/district', 'middleware' => ['auth', 'role:superadministrator|admin_sekolah']], function() {
     $class          = 'Bantenprov\WilayahIndonesia\Http\Controllers\DistrictController';
     $name           = 'wilayah-indonesia.district';
     $controllers    = (object) [
@@ -106,18 +106,18 @@ Route::group(['prefix' => 'api/wilayah-indonesia/district', 'middleware' => ['we
         'destroy'   => $class.'@destroy',
     ];
 
-    Route::get('/',                 $controllers->index)->name($name.'.index');
-    Route::get('/get',              $controllers->get)->name($name.'.get');
-    Route::get('/get/by-city/{id}', $controllers->getByCity)->name($name.'.get-by-city');
-    Route::get('/create',           $controllers->create)->name($name.'.create');
-    Route::get('/{id}',             $controllers->show)->name($name.'.show');
-    Route::post('/',                $controllers->store)->name($name.'.store');
-    Route::get('/{id}/edit',        $controllers->edit)->name($name.'.edit');
-    Route::put('/{id}',             $controllers->update)->name($name.'.update');
-    Route::delete('/{id}',          $controllers->destroy)->name($name.'.destroy');
+    Route::get('/',                 $controllers->index)->name($name.'.index')->middleware(['role:superadministrator']);
+    Route::get('/get',              $controllers->get)->name($name.'.get')->middleware(['role:superadministrator']);
+    Route::get('/get/by-city/{id}', $controllers->getByCity)->name($name.'.get-by-city')->middleware(['role:superadministrator|admin_sekolah']);
+    Route::get('/create',           $controllers->create)->name($name.'.create')->middleware(['role:superadministrator']);
+    Route::get('/{id}',             $controllers->show)->name($name.'.show')->middleware(['role:superadministrator']);
+    Route::post('/',                $controllers->store)->name($name.'.store')->middleware(['role:superadministrator']);
+    Route::get('/{id}/edit',        $controllers->edit)->name($name.'.edit')->middleware(['role:superadministrator']);
+    Route::put('/{id}',             $controllers->update)->name($name.'.update')->middleware(['role:superadministrator']);
+    Route::delete('/{id}',          $controllers->destroy)->name($name.'.destroy')->middleware(['role:superadministrator']);
 });
 
-Route::group(['prefix' => 'api/wilayah-indonesia/village', 'middleware' => ['web']], function() {
+Route::group(['prefix' => 'api/wilayah-indonesia/village', 'middleware' => ['auth', 'role:superadministrator|admin_sekolah']], function() {
     $class          = 'Bantenprov\WilayahIndonesia\Http\Controllers\VillageController';
     $name           = 'wilayah-indonesia.village';
     $controllers    = (object) [
@@ -132,13 +132,13 @@ Route::group(['prefix' => 'api/wilayah-indonesia/village', 'middleware' => ['web
         'destroy'       => $class.'@destroy',
     ];
 
-    Route::get('/',                     $controllers->index)->name($name.'.index');
-    Route::get('/get',                  $controllers->get)->name($name.'.get');
-    Route::get('/get/by-district/{id}', $controllers->getByDistrict)->name($name.'.get-by-district');
-    Route::get('/create',               $controllers->create)->name($name.'.create');
-    Route::get('/{id}',                 $controllers->show)->name($name.'.show');
-    Route::post('/',                    $controllers->store)->name($name.'.store');
-    Route::get('/{id}/edit',            $controllers->edit)->name($name.'.edit');
-    Route::put('/{id}',                 $controllers->update)->name($name.'.update');
-    Route::delete('/{id}',              $controllers->destroy)->name($name.'.destroy');
+    Route::get('/',                     $controllers->index)->name($name.'.index')->middleware(['role:superadministrator']);
+    Route::get('/get',                  $controllers->get)->name($name.'.get')->middleware(['role:superadministrator']);
+    Route::get('/get/by-district/{id}', $controllers->getByDistrict)->name($name.'.get-by-district')->middleware(['role:superadministrator|admin_sekolah']);
+    Route::get('/create',               $controllers->create)->name($name.'.create')->middleware(['role:superadministrator']);
+    Route::get('/{id}',                 $controllers->show)->name($name.'.show')->middleware(['role:superadministrator']);
+    Route::post('/',                    $controllers->store)->name($name.'.store')->middleware(['role:superadministrator']);
+    Route::get('/{id}/edit',            $controllers->edit)->name($name.'.edit')->middleware(['role:superadministrator']);
+    Route::put('/{id}',                 $controllers->update)->name($name.'.update')->middleware(['role:superadministrator']);
+    Route::delete('/{id}',              $controllers->destroy)->name($name.'.destroy')->middleware(['role:superadministrator']);
 });
